refactor: error handling and config validations
