+++ packages/config/src/bundle-config.ts
@@ -1,15 +1,9 @@
-import { ConfigError, ConfigNotFoundError } from '@pandacss/error'
 import { logger } from '@pandacss/logger'
+import { PandaError } from '@pandacss/shared'
 import type { Config } from '@pandacss/types'
 import { bundleNRequire } from 'bundle-n-require'
 import { findConfig } from './find-config'
-import type { ConfigFileOptions } from './types'
-
-export interface BundleConfigResult<T = Config> {
-  config: T
-  dependencies: string[]
-  path: string
-}
+import type { BundleConfigResult, ConfigFileOptions } from './types'
 
 export async function bundle<T = Config>(filepath: string, cwd: string) {
   const { mod: config, dependencies } = await bundleNRequire(filepath, {
@@ -27,17 +21,12 @@ export async function bundleConfig(options: ConfigFileOptions): Promise<BundleCo
 
   const filePath = findConfig({ cwd, file })
 
-  if (!filePath) {
-    throw new ConfigNotFoundError()
-  }
-
   logger.debug('config:path', filePath)
 
   const result = await bundle(filePath, cwd)
 
-  // TODO: Validate config shape
   if (typeof result.config !== 'object') {
-    throw new ConfigError(`💥 Config must export or return an object.`)
+    throw new PandaError('CONFIG_ERROR', `💥 Config must export or return an object.`)
   }
 
   result.config.outdir ??= 'styled-system'
