+++ packages/config/src/validation/get-final-paths.ts
@@ -1,0 +1,11 @@
+export const getFinalPaths = (paths: Set<string>) => {
+  paths.forEach((path) => {
+    paths.forEach((potentialExtension) => {
+      if (potentialExtension.startsWith(path + '.')) {
+        paths.delete(path)
+      }
+    })
+  })
+
+  return paths
+}
