+++ packages/core/src/style-decoder.ts
@@ -1,4 +1,13 @@
-import { deepSet, esc, getOrCreateSet, isImportant, markImportant, toHash, withoutImportant } from '@pandacss/shared'
+import {
+  PandaError,
+  deepSet,
+  esc,
+  getOrCreateSet,
+  isImportant,
+  markImportant,
+  toHash,
+  withoutImportant,
+} from '@pandacss/shared'
 import type {
   AtomicStyleResult,
   Dict,
@@ -296,7 +305,7 @@ export class StyleDecoder {
     const recipeConfig = this.context.recipes.getConfigOrThrow(recipeName)
 
     if (!Recipes.isSlotRecipeConfig(recipeConfig)) {
-      throw new Error(`Recipe "${recipeName}" is not a slot recipe`)
+      throw new PandaError('UNKNOWN_RECIPE', `Recipe "${recipeName}" is not a slot recipe`)
     }
 
     const base: Dict = {}
