--- packages/error/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "@pandacss/error",
-  "version": "0.28.0",
-  "description": "Common error messages for css panda",
-  "author": "Segun Adebayo <joseshegs@gmail.com>",
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
-  "sideEffects": false,
-  "homepage": "https://panda-css.com",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/chakra-ui/panda.git",
-    "directory": "packages/error"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "exports": {
-    ".": {
-      "source": "./src/index.ts",
-      "types": "./dist/index.d.ts",
-      "require": "./dist/index.js",
-      "import": {
-        "types": "./dist/index.d.mts",
-        "default": "./dist/index.mjs"
-      }
-    },
-    "./package.json": "./package.json"
-  },
-  "scripts": {
-    "build": "tsup src/index.ts --format=esm,cjs --dts",
-    "build-fast": "tsup src/index.ts --format=esm,cjs --no-dts",
-    "dev": "pnpm build-fast --watch"
-  },
-  "files": [
-    "dist"
-  ]
-}
