--- packages/error/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "include": ["src"]
-}
