+++ packages/shared/src/index.ts
@@ -12,6 +12,7 @@ export * from './css-unit'
 export * from './css-var'
 export * from './deep-set'
 export * from './entries'
+export * from './error'
 export * from './esc'
 export * from './flatten'
 export * from './get-or-create-set'
